+++ CHANGES.md
@@ -1,6 +1,8 @@
 Changes 1.1.1-prerelease:
 * fix: properly use the mutex absolute time argument, timedlock handling and mutex destruction in the ESDM RPC client lib
 
+* fix: race condition in worker thread execution
+
 Changes 1.1.0:
 * fix: name of leancrypto DRNG